refactor(envsecrets): rename internal/secrets → internal/envsecrets (#341)
Two "secrets" packages under internal/ was a confusing naming collision:

- internal/secrets/ — Docker/Kubernetes env-var reader (18 LoC, one helper)
- internal/sandbox/secrets/ — per-app vault + proxy socket

Renamed the env-var reader to internal/envsecrets to remove the ambiguity;
both now map to distinct import identifiers at every call-site. Added a
package doc comment that points at sandbox/secrets to keep the distinction
obvious to future readers.

No behavior change. 4 call-sites updated (cmd/alf-daemon/{main,config}.go,
cmd/extract-video/main.go, vault import alias kept as 'vault').
