+++ cmd/extract-video/main.go
@@ -29,7 +29,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/alamparelli/alf/internal/secrets"
+	"github.com/alamparelli/alf/internal/envsecrets"
 	"github.com/alamparelli/alf/internal/voice"
 )
 
@@ -193,7 +193,7 @@ func collectFiles(prefix string, n int) ([]string, error) {
 
 func transcribeAudio(videoPath string) (string, string) {
 	whisperURL := os.Getenv("WHISPER_URL")
-	whisperSecret := secrets.ReadSecret("WHISPER_SHARED_SECRET")
+	whisperSecret := envsecrets.ReadSecret("WHISPER_SHARED_SECRET")
 	if whisperURL == "" || whisperSecret == "" {
 		return "", ""
 	}
